Add files via upload
In many research projects (for example, in hardware projects) to perform a specific task (for example, changing the multiplication operation, obtaining the maximum and implementing different functions based on a new hardware), it is necessary to be able to convert the trained weights into integer numbers. and converted in a specific range of numbers (for example 8-bit numbers) so that the process of converting these weights into binary numbers in that specific range is easier. A special scale is written and after applying the desired operation on these numbers or this array, we return the array numbers to their previous state using the Dequantize function.
For example, in the convolution layer, to multiply the input by weight, we first quantize both of these arrays based on a scale, and then multiply based on the quantized numbers and then dequantize the desired output. And according to the test that has been done, the work of the functions is done correctly.